pyproject.toml
[tool.poetry]
authors = [
  "Bishwarup B <write2bishwarup@gmail.com>",
  "Satheesh K <satheeshkatipomu@gmail.com>",
  "Aman G <amangupta0044@gmail.com>",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Scientific/Engineering :: Image Recognition",
]
description = "Utilities for vision related tasks"
homepage = "https://github.com/hashtagml/optical"
include = ["MANIFEST.in", "CONTRIBUTING.md", "LICENSE"]
keywords = ["computer vision", "deep learning", "object detection"]
license = "MIT"
name = "optical"
readme = "README.md"
version = "0.0.2"

[tool.poetry.dependencies]
python=">=3.8,<3.11"
pandas = "^1.4.2"
scipy = "~1.8"
scikit-learn = "~1"
PyYAML = "^6.0"
opencv-python-headless = "^4.5.5"
imagesize = "^1.3.0"
Pillow = "^9.1.0"
lxml = "^4.8.0"
tqdm = "^4.64.0"
matplotlib = "^3.5.2"
altair = "^4.2.0"
scikit-multilearn = "^0.2.0"
mediapy = "^1.0.3"
tensorflow-cpu = {version = "^2.8.0", optional = true}

[tool.poetry.extras]
tensorflow = ["tensorflow-cpu"]

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
autoflake = "^1.4"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
kaggle = "^1.5.12"
tox = "^3.25.0"
Sphinx = "^4.5.0"
sphinx-panels = "^0.6.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-copybutton = "^0.5.0"
bump2version = "^1.0.1"
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
  
[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
include = '\.pyi?$'
line-length = 119
target-version = ['py38', 'py39', 'py310']

[tool.pytest.ini_options]
filterwarnings = [
  "error",
  "ignore::UserWarning",
  "ignore::DeprecationWarning",
  'ignore:function ham\(\) is deprecated:DeprecationWarning',
] # note the use of single quote below to denote "raw" strings in TOML
